# artflow/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APIX",
    "description": "APIX AI API for Telegram auth, content generation, billing, prompt library, feed, and provider webhooks.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "miniapp"
        ],
        "summary": "Get History",
        "description": "Paginated generations for the current user, newest first.",
        "operationId": "get_history_api_v1_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-telegram-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          },
          {
            "name": "x-web-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Web-Auth-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenerationOut"
                  },
                  "title": "Response Get History Api V1 History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/web/history": {
      "get": {
        "tags": [
          "web"
        ],
        "summary": "History",
        "operationId": "history_api_web_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Dev-Tg-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Tg-Id"
            }
          },
          {
            "name": "X-Web-Auth-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Web-Auth-Token"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "gen_type": {
            "type": "string",
            "title": "Gen Type"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "credits_spent": {
            "type": "number",
            "title": "Credits Spent"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "is_public_feed": {
            "type": "boolean",
            "title": "Is Public Feed",
            "default": false
          },
          "is_prompt_library": {
            "type": "boolean",
            "title": "Is Prompt Library",
            "default": false
          },
          "result_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Result Urls",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "model",
          "gen_type",
          "prompt",
          "status",
          "result_url",
          "credits_spent",
          "created_at"
        ],
        "title": "GenerationOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}